+++ gframe/menu_handler.cpp
@@ -38,8 +38,8 @@ static void UpdateDeck() {
 	const auto totsize = deck.main.size() + deck.extra.size() + deck.side.size();
 	if(totsize > max_deck_size)
 		return;
-	BufferIO::Write<uint32_t>(pdeck, deck.main.size() + deck.extra.size());
-	BufferIO::Write<uint32_t>(pdeck, deck.side.size());
+	BufferIO::Write<uint32_t>(pdeck, static_cast<uint32_t>(deck.main.size() + deck.extra.size()));
+	BufferIO::Write<uint32_t>(pdeck, static_cast<uint32_t>(deck.side.size()));
 	for(const auto& pcard : deck.main)
 		BufferIO::Write<uint32_t>(pdeck, pcard->code);
 	for(const auto& pcard : deck.extra)
