+++ gframe/replay.cpp
@@ -21,7 +21,7 @@ void Replay::BeginRecord(bool write, epro::path_string name) {
 }
 void Replay::WritePacket(const CoreUtils::Packet& p) {
 	Write<uint8_t>(p.message, false);
-	Write<uint32_t>(p.buff_size(), false);
+	Write<uint32_t>(static_cast<uint32_t>(p.buff_size()), false);
 	WriteData(p.data(), p.buff_size());
 }
 bool Replay::IsStreamedReplay() {
@@ -67,7 +67,7 @@ void Replay::EndRecord(size_t size) {
 		fclose(fp);
 		fp = nullptr;
 	}
-	pheader.base.datasize = replay_data.size() - sizeof(ExtendedReplayHeader);
+	pheader.base.datasize = static_cast<uint32_t>(replay_data.size() - sizeof(ExtendedReplayHeader));
 	pheader.base.flag |= REPLAY_COMPRESSED;
 	size_t propsize = 5;
 	auto comp_size = size;
@@ -220,7 +220,7 @@ void Replay::ParseParams() {
 	else
 		params.duel_flags = Read<uint32_t>();
 	if(pheader.base.flag & REPLAY_SINGLE_MODE && pheader.base.id == REPLAY_YRP1) {
-		size_t slen = Read<uint16_t>();
+		uint16_t slen = Read<uint16_t>();
 		scriptname.resize(slen);
 		ReadData(&scriptname[0], slen);
 	}
@@ -380,7 +380,7 @@ bool Replay::ParseResponses() {
 	return !responses.empty();
 }
 
-bool ExtendedReplayHeader::ParseReplayHeader(const void* data, uint32_t input_len, ExtendedReplayHeader& header, uint32_t* header_length) {
+bool ExtendedReplayHeader::ParseReplayHeader(const void* data, size_t input_len, ExtendedReplayHeader& header, uint32_t* header_length) {
 	if(input_len < sizeof(ReplayHeader))
 		return false;
 	ExtendedReplayHeader ret{};
